requirements.txt
elasticsearch==6.1.1
Django>=2.0
django-cors-headers==2.1.0
djangorestframework==3.7.7
PyMysql==0.8.0
django-redis==4.8.0
django-filter==1.1.0
django-rest-framework-mongoengine==3.3.1
django-rest-swagger==2.1.2

mongoengine==0.15.0
pdfminer3k==1.3.1
flytrap-base==0.0.2
flytrap-auth==0.0.5
flytrap-comments==0.0.1
Pillow==5.0.0
pytest-django==3.1.2
